Rename build utils
